Reduce complexity of Install/Uninstall gpm action
This PR reduces the complexity of checking the visibility status of the GPM action "Install/Uninstall". This action field is only hidden on those plugins with a `enabled.type: hidden` field and shown otherwise. This PR especially accounts for plugins which have an `enabled: true` property, but their `enabled` option in the blueprint is located elsewhere (e.g. wrapped in a section field).