Fix for JS error in Safari private browsing, #272
The problem is that Safari in private mode blocks Local Storage and
Session Storage, setting a quota of 0 so all attempts to save something
generate an error and all the subsequent JS breaks.

If setItem does not work, we set an empty prototype implementation that
does not generate an error, to avoid having to add multiple checks in
the code.

This only affects being able to restore the pages expanded tree
(assuming a page was expanded prior to reloading the page) when in
Safari private browsing.