Fix #323, regression introduced in b7caab8677b00c7ed2e2b43adf3cb2d802d0fad0
@w00fz I partially reverted b7caab8677b00c7ed2e2b43adf3cb2d802d0fad0, I
tested and it works with the markdown editor too. What was the reason
you changed it?

The problem introduced was the `data-collection-template` element was
changed from <script> to <ul> and then the template item is picked up
as a list item in the list, with '*' as key, instead of being ignored.