Restore file inputs functionality. Process the form via javascript only if no file inputs are found.
The file inputs are not found in Pages, so in that case, the JS form
processing still kicks in as usual. In plugins and themes blueprints on
the other hand, if there is a file input we avoid JS processing, since
we also do not actually need this part of code in those forms.