Use scope also in admin registration / login form
BC issue: if using Admin `develop/ 1.1` with Form != latest develop,
the Admin-side registration form (the one that pops up when Admin is
installed but no accounts are found) will not work. Not something users
would stumble upon: only appears when Admin  1.1 it'll be possible to
use config-title@: theme.antimatter.some_config to set the title
dynamically based on a theme config option is installed but it's not
yet configured, and Form is not up to date. When 1.1 will be released,
also the Form plugin will have a newer update that will be installed
automatically.