Fix Ctrl+S does not save changes to a page
Press Ctrl+S when editing a page does trigger a save prompt, however,
the content is not saved.

This was broken because Remodal changed their DOM structure in commit
461d5ec307a3214e0d2c1427b613178236ed9293, which caused the
"remodal-overlay" not to be a parent of the "Continue" button.

This commit updated the selector to rely on the DOM structure we
control, so that hopefully it will not break again if Remodal changes.