Admin panel cleaner look for tabs, theme & plugin
Improvement to tab parts of admin interface.

Ideas:
* One idea was to inject extra HTML code higher up in the twig structure into the field tab interface. However this would require the storage of context to be passed to each twig instance, complicating the structure
* Second idea was to inject this on JavaScript side. However that attributes to more DOM traversal and manipulation, and performance is better to render this server-side.
* Third idea was to have special config parameters added to the actual tab interface. However this would not make the tab instance

Proposed:
Custom field type for themes and plugins. Conditional context awareness if this is a plugin or a theme. The theme or plugin twig variables only exist in the blueprint context is appropriate. If in theme, the plugin twig variable is null. If in plugin, the theme twig variable is null. We can use this to check the field appropriately and output a custom message so developers can know what field type to use when.