Fix for user without config_system permissions (#1540)
When a user doesn't have admin.configuration_system permissions, rather than opening the next tab (site) it just blanks out without a tab open. This simple fix let's it check for system permission and if not add site. Ideally we should check in order:

if not system use site
if not site use media
if not media use info

but I don't see that being required too much as Info is the most likely one for a user to have if not system (since its just the basic site information)